--- rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalizer/package.scala
@@ -1,3 +1,0 @@
-package coop.rchain.rholang.interpreter.compiler
-
-package object normalizer {}
