+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalizer/processes/PBundleNormalizer.scala
@@ -1,8 +1,8 @@
-package coop.rchain.rholang.interpreter.compiler.normalizer
+package coop.rchain.rholang.interpreter.compiler.normalizer.processes
 
 import cats.syntax.all._
 import cats.effect.Sync
-import coop.rchain.models.{Bundle, BundleOps, Par}
+import coop.rchain.models.{BundleOps, Par}
 import coop.rchain.models.rholang.implicits._
 import coop.rchain.rholang.interpreter.compiler.ProcNormalizeMatcher.normalizeMatch
 import coop.rchain.rholang.interpreter.compiler.{
@@ -13,7 +13,6 @@ import coop.rchain.rholang.interpreter.compiler.{
 }
 import coop.rchain.rholang.interpreter.errors.UnexpectedBundleContent
 import coop.rchain.rholang.ast.rholang_mercury.Absyn.{
-  Bundle => AbsynBundle,
   BundleEquiv,
   BundleRead,
   BundleReadWrite,
