+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalizer/processes/PCollectNormalizer.scala
@@ -1,4 +1,4 @@
-package coop.rchain.rholang.interpreter.compiler.normalizer
+package coop.rchain.rholang.interpreter.compiler.normalizer.processes
 
 import cats.syntax.all._
 import cats.effect.Sync
@@ -10,6 +10,7 @@ import coop.rchain.rholang.interpreter.compiler.{
   ProcVisitOutputs
 }
 import coop.rchain.rholang.ast.rholang_mercury.Absyn.PCollect
+import coop.rchain.rholang.interpreter.compiler.normalizer.CollectionNormalizeMatcher
 
 object PCollectNormalizer {
   def normalize[F[_]: Sync](p: PCollect, input: ProcVisitInputs)(
