+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalizer/processes/PEvalNormalizer.scala
@@ -1,4 +1,4 @@
-package coop.rchain.rholang.interpreter.compiler.normalizer
+package coop.rchain.rholang.interpreter.compiler.normalizer.processes
 
 import cats.syntax.all._
 import cats.effect.Sync
@@ -6,6 +6,7 @@ import coop.rchain.models.Par
 import coop.rchain.models.rholang.implicits._
 import coop.rchain.rholang.interpreter.compiler.{NameVisitInputs, ProcVisitInputs, ProcVisitOutputs}
 import coop.rchain.rholang.ast.rholang_mercury.Absyn.PEval
+import coop.rchain.rholang.interpreter.compiler.normalizer.NameNormalizeMatcher
 
 object PEvalNormalizer {
   def normalize[F[_]: Sync](p: PEval, input: ProcVisitInputs)(
