+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalizer/processes/PGroundNormalizer.scala
@@ -1,11 +1,11 @@
-package coop.rchain.rholang.interpreter.compiler.normalizer
+package coop.rchain.rholang.interpreter.compiler.normalizer.processes
 
 import cats.syntax.all._
 import cats.effect.Sync
-import coop.rchain.models.Par
 import coop.rchain.models.rholang.implicits._
 import coop.rchain.rholang.interpreter.compiler.{ProcVisitInputs, ProcVisitOutputs}
 import coop.rchain.rholang.ast.rholang_mercury.Absyn.PGround
+import coop.rchain.rholang.interpreter.compiler.normalizer.GroundNormalizeMatcher
 
 object PGroundNormalizer {
   def normalize[F[_]: Sync](p: PGround, input: ProcVisitInputs): F[ProcVisitOutputs] =
