+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalizer/processes/PSimpleTypeNormalizer.scala
@@ -1,4 +1,4 @@
-package coop.rchain.rholang.interpreter.compiler.normalizer
+package coop.rchain.rholang.interpreter.compiler.normalizer.processes
 
 import cats.syntax.all._
 import cats.effect.Sync
